packages/try/try-nicegui/pyproject.toml
[project]
name = "try-nicegui"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Henry.Huang", email = "hhglory@outlook.com" }
]

[tool.poetry]
name = "try-nicegui"
version = "0.1.0"
description = ""
authors = ["Henry.Huang <hhglory@outlook.com>"]
readme = "README.md"
packages = [{ include = "try_nicegui", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
click = "^8.1.3"
pystray = "^0.19.4"
loguru = "^0.7.0"
pillow = "^8.3.2"
pywebview = "^4.1"
typer = { extras = ["all"], version = "^0.9.0" }
nicegui = "^1.4.15"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.12.0"
black = "^23.3.0"
pytest = "^7.3.1"
cx-freeze = "^6.15.3"
nuitka = "^1.9.5"
libpython = "^0.2" # for nuitka

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
